pyproject.toml
[project]
name = "untitled_project"
version = "0.1.0"
description = "untitled_project using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.119.0,<1.0.0",
    "PyGithub>=2.1.1"
]

[project.scripts]
untitled_project = "untitled_project.main:run"
run_crew = "untitled_project.main:run"
train = "untitled_project.main:train"
replay = "untitled_project.main:replay"
test = "untitled_project.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
